Move debug info types to types dir
